Update app.py
Major overhaul of the main syslog/data collection script.  
Testing with the PfSense Plus 21.05 releases and the PfSense 2.5.x Community edition releases have uncovered a bug the was pre-existing in the 2.4.x versions regarding the formatting of logs in BSD format versus standard RFC syslog is still present in the 21.05 release even though the RFC logging option has now been added.  
Due to the system needing to work in a mixed 21.05 and 2.5.x environment and changes within the PfSense team and dev cycle making it likely this issue will not be resolved soon a new method was needed.  
The following code uses the previous SSH access to pull logs directly from the underlying FreeBSD file system, bypassing these issues.  
With the system now relying purely on SSH access to add a new instance to the system an admin will need to add the machine details to the DB, at which point the system will autmatically poll it, versus the previous method where once the system started receiving Syslog messages from an instance it would automatically add it to the system.
However this new implementation bring with it greater flexibility to access a broader range of logs from the underlying /var/logs folder, and no longer requires port forwarding from the host container